requirements.txt
ocrmac
pillow
pynput
pyobjc-framework-Quartz
